COMMERCE-2475 Does not deserve to be a public api.
