+++ modules/apps/dispatch/dispatch-api/bnd.bnd
@@ -2,7 +2,6 @@ Bundle-Name: Liferay Dispatch API
 Bundle-SymbolicName: com.liferay.dispatch.api
 Bundle-Version: 1.0.0
 Export-Package:\
-	com.liferay.dispatch,\
 	com.liferay.dispatch.constants,\
 	com.liferay.dispatch.exception,\
 	com.liferay.dispatch.model,\
