--- modules/apps/dispatch/dispatch-api/src/main/resources/com/liferay/dispatch/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
